requirements.txt
python-dotenv
aiogram>=3.0.0
flask
flask-sqlalchemy
